refactor(pacakge): package structure
