+++ package/resources/constructs/test/frontend-stack.json
