+++ package/resources/internal-constructs/base-pipeline.ts
