+++ package/resources/internal-constructs/credential-store.spec.ts
