+++ package/resources/internal-constructs/credential-store.ts
