+++ package/resources/internal-constructs/frontend-cdn-pipeline.spec.ts
