+++ package/resources/internal-constructs/slack-subscription.spec.ts
