+++ package/resources/internal-constructs/slack-subscription.ts
