+++ package/resources/internal-constructs/test/multi-env-approval-pipeline.spec.json
