+++ package/resources/internal-constructs/test/multi-env-pipeline-no-approval.spec.json
