+++ package/resources/internal-constructs/test/single-env-pipeline.spec.json
