+++ package/resources/internal-constructs/web-distribution.ts
